Update ObjectQueueExecutor.java
This chang is to fix an issue with a catched exception that is caused by a failing database transaction (for example not explicitly committed objects that are in the state 'Autocommitted'). Currently the process queue is only creating an info log but it is obviously an error. Changed it to error. Using the existing context for the setErrormessageAndCommit method is causing the changes made by his action to be rolled back with the entire transaction. I created a new system context to run this action. If you don't do that, the QueuedAction will remain in running forever, even if it is already terminated.